serverless/requirements.txt
annotated-types==0.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
dnspython==2.7.0
idna==3.10
pydantic==2.11.1
pydantic_core==2.33.0
pymongo==4.11.3
python-dotenv==1.1.0
requests==2.32.3
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
